Moved HTTP sources to dedicated module
