+++ cerver/__init__.py
@@ -2,15 +2,4 @@
 from .cerver import *
 from .version import *
 
-# HTTP
-from .alg import *
-from .content import *
-from .headers import *
-from .http import *
-from .query import *
-from .request import *
-from .response import *
-from .route import *
-from .status import *
-
 __version__ = "0.5.1"
