+++ cerver/http/content.py
@@ -1,6 +1,6 @@
 from ctypes import c_int, c_char_p, c_bool
 
-from .lib import lib
+from ..lib import lib
 
 ContentType = c_int
